split classes portable
