+++ src/serialization/portable/ClassDefinition.ts
